refactor(/fixture): Fixture들의 패키지를 util -> fixture로 변경
