+++ src/test/java/roomescape/reservation/repository/JdbcReservationRepositoryTest.java
@@ -24,10 +24,10 @@ import roomescape.testutil.JdbcRepositoryTest;
 import roomescape.theme.model.Theme;
 import roomescape.theme.repository.JdbcThemeRepository;
 import roomescape.theme.repository.ThemeRepository;
-import roomescape.util.MemberFixture;
-import roomescape.util.ReservationFixture;
-import roomescape.util.ReservationTimeFixture;
-import roomescape.util.ThemeFixture;
+import roomescape.fixture.MemberFixture;
+import roomescape.fixture.ReservationFixture;
+import roomescape.fixture.ReservationTimeFixture;
+import roomescape.fixture.ThemeFixture;
 
 @JdbcRepositoryTest
 class JdbcReservationRepositoryTest {
